Remove newline from the test text (hello_world.txt)
The newline character '\n' is stored to a file differently depending on the platform. Especially, Windows uses two characters ('\r\n'), so the size of string and the size of file are not same, which makes the size comparison test fail. We can make this test platform independent later but I just quickly removed the newline character for now because it has nothing to do with the functionality that we want to test here.