fixed tutorialDominoes-Finished - this did not work
I tried the tutorial but it did not work. I could never add
a new domino. There were three issues in the code.

a) The new domino was never added to the world.
b) The collision detection was always detecting a collision because
   the new domino is in contact with the floor.
c) The tutorial description on the website is not related to
   the code for the collision detection.

I fixed a) and b) and added some comments with this commit.