Import dependency targets when loading DARTConfig.cmake (#1209)
Currently, the transitive dependencies of DART are propagated by using CMake variables. For example, all the dependency libraries are stored in `DART_LIBRARIES`, which will be defined by loading `DARTConfig.cmake` through calling `find_package(DART)` in a downstream project. 

However, this approach can fail to load the dependency. The items in `DART_LIBRARIES` can be either of the path to a library file or a target. Path items work fine (as long as the path is correct), but not targets because the downstream project don't know how to import the target only from the target name. This PR fixes it by letting `DARTConfig.cmake` loading the dependency targets by calling `find_packages()` for them. This approach should work as long as all the upstream dependencies loading relevant targets subsequently.

In this PR, the new change is only applied for loading `urdfdom` to quickly fix #1200, but this should be applied to all other transitive dependencies in future PRs.

***

**Before merging a pull request**

- [x] Set version target by selecting a milestone on the right side