Add a RealTimeWorldNode for displaying simulations at real-time speeds (#1216)
I recently needed to make a real-time simulation display on top of dartsim, so I decided to clean it up and add it into the main repo since it's a frequently requested feature (e.g. #1204).

It's a simple wrapper on top of the `dart::gui::osg::WorldNode` class. Users can opt for either the old `WorldNode` or the new `RealTimeWorldNode` depending on how they want to control the speed of the simulation.

I also added a line `ImGui::CreateContext();` because ImGui was crashing without that.

I know #1215 has already been opened, so I can retarget this at a `6.8` release (or we can wait until 6.7.1 since this won't impact API or ABI) instead if that would be preferred.

***

**Before creating a pull request**

- [x] Document new methods and classes
- [x] Format new code files using `clang-format`

**Before merging a pull request**

- [x] Set version target by selecting a milestone on the right side
- [x] Summarize this change in `CHANGELOG.md`
- [ ] (N/A) Add unit test(s) for this change